+++ test/test.js
@@ -1,1 +1,3 @@
-require('./app.js');
+require('./sample.js');
+require('./submit.js');
+require('./upload.js');
